Factor out devQuery and _formatError from createGraphqlMiddleware
