+++ .changeset/143327d0/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/server", "type": "patch" }], "dependents": [] }
